pyproject.toml
[project]
name = "trend-line-strat"
version = "0.1.0"
description = "Trendline breakout detection and backtesting framework"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["trading", "trendline", "backtesting", "optuna", "crypto"]
authors = [ { name = "Maurice Colling", email = "collingmaurice@web.de" } ]
dependencies = [
    "pandas>=2.1",
    "numpy>=1.24",
    "scipy>=1.11",
    "requests>=2.31",
    "python-binance>=1.0",
    "optuna>=3.5",
    "pysqlite3-binary>=0.5",
    "typer>=0.12",
    "rich>=13.7"
]

[project.optional-dependencies]
test = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "jsonschema>=4.19"
]
dev = [
    "ruff>=0.5",
    "black>=24.8",
    "mypy>=1.10"
]

[project.scripts]
trend-line-strat = "src.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
addopts = "-ra --strict-config --strict-markers --disable-warnings --cov=src --cov-report=term-missing"
testpaths = ["tests"]
xfail_strict = true

[tool.black]
line-length = 100
target-version = ["py311"]
include = "\\.pyi?$"

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "W", "UP", "S", "C4", "DTZ"]
ignore = ["S101"]
fix = true

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_return_any = true
warn_unused_configs = true
plugins = []
